[SQL] Move some classes into packages that are more appropriate.
