+++ sql/core/src/main/scala/org/apache/spark/sql/execution/pythonUdfs.scala
@@ -20,6 +20,7 @@ package org.apache.spark.sql.execution
 import java.util.{List => JList, Map => JMap}
 
 import org.apache.spark.rdd.RDD
+import org.apache.spark.sql.catalyst.util.DateUtils
 
 import scala.collection.JavaConversions._
 import scala.collection.JavaConverters._
