+++ sql/core/src/main/scala/org/apache/spark/sql/json/JsonRDD.scala
@@ -19,6 +19,8 @@ package org.apache.spark.sql.json
 
 import java.sql.Timestamp
 
+import org.apache.spark.sql.catalyst.util.DateUtils
+
 import scala.collection.Map
 import scala.collection.convert.Wrappers.{JMapWrapper, JListWrapper}
