+++ sql/core/src/test/scala/org/apache/spark/sql/json/JsonSuite.scala
@@ -21,6 +21,7 @@ import java.io.StringWriter
 import java.sql.{Date, Timestamp}
 
 import com.fasterxml.jackson.core.JsonFactory
+import org.apache.spark.sql.catalyst.util.DateUtils
 import org.scalactic.Tolerance._
 
 import org.apache.spark.sql.TestData._
